+++ components/camel-beanstalk/src/test/java/org/apache/camel/component/beanstalk/integration/BuryProducerIntegrationIT.java
@@ -28,6 +28,7 @@ import org.apache.camel.component.beanstalk.Headers;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -35,7 +36,9 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class BuryProducerIntegrationTest extends BeanstalkCamelTestSupport {
+@EnabledIfSystemProperty(named = "beanstalk.host", matches = ".*",
+                         disabledReason = "Requires a beanstalk server running")
+public class BuryProducerIntegrationIT extends BeanstalkCamelITSupport {
     @EndpointInject("mock:result")
     protected MockEndpoint resultEndpoint;
