+++ components/camel-beanstalk/src/test/java/org/apache/camel/component/beanstalk/integration/TouchProducerIntegrationIT.java
@@ -26,7 +26,9 @@ import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.beanstalk.Headers;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -34,15 +36,18 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class DeleteProducerIntegrationTest extends BeanstalkCamelTestSupport {
+@EnabledIfSystemProperty(named = "beanstalk.host", matches = ".*",
+                         disabledReason = "Requires a beanstalk server running")
+public class TouchProducerIntegrationIT extends BeanstalkCamelITSupport {
     @EndpointInject("mock:result")
     protected MockEndpoint resultEndpoint;
 
     @Produce("direct:start")
     protected ProducerTemplate direct;
 
+    @Disabled("requires reserve - touch sequence")
     @Test
-    void testDelete() throws InterruptedException, IOException {
+    void testBury() throws InterruptedException, IOException {
         long jobId = writer.put(0, 0, 5, new byte[0]);
         assertTrue(jobId > 0, "Valid Job Id");
 
@@ -57,8 +62,12 @@ public class DeleteProducerIntegrationTest extends BeanstalkCamelTestSupport {
         assertNotNull(messageJobId, "Job ID in message");
         assertEquals(jobId, messageJobId.longValue(), "Message Job ID equals");
 
-        final Job job = reader.peek(jobId);
-        assertNull(job, "Job has been deleted");
+        final Job job = reader.reserve(0);
+        assertNull(job, "Beanstalk client has no message");
+
+        final Job buried = reader.peekBuried();
+        assertNotNull(buried, "Job in buried");
+        assertEquals(jobId, buried.getJobId(), "Buried job id");
     }
 
     @Test
@@ -73,7 +82,7 @@ public class DeleteProducerIntegrationTest extends BeanstalkCamelTestSupport {
         return new RouteBuilder() {
             @Override
             public void configure() {
-                from("direct:start").to("beanstalk:" + tubeName + "?command=delete").to("mock:result");
+                from("direct:start").to("beanstalk:" + tubeName + "?command=touch").to("mock:result");
             }
         };
     }
